chpore: remove magic numbers
Previously there was a lot of 0.001 and 0.002 sprinkled across the
source code. This commit introduces a new constant EPS which is
reasonably small to not change anything in the mechanical side,
but to please on the various instances of floating point weirdness.
Subsequently, all instances of the magic numbers are replaced by EPS or
multiples of EPS.